Update 12.05 Add PlayerControl component Update backend url to environment variable Add new pause photo Add yarn.lock Add package-lock.json
