+++ src/components/playerControl/PlayerControl.module.css
@@ -1,0 +1,27 @@
+.play-container {
+    justify-content: center;
+}
+.play-group {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+}
+#play-button {
+    width: 40px;
+    height: 40px;
+}
+.next-track {
+    display: flex;
+    justify-content: left;
+}
+#next-button {
+    width: 2em;
+}
+.prev-track {
+    display: flex;
+    justify-content: right;
+}
+#prev-button {
+    width: 2em;
+}
